requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
python-magic==0.4.27
celery==5.3.6
redis==5.0.1
asyncpg==0.29.0
alembic==1.13.1 